+++ mapca/__init__.py
@@ -3,26 +3,13 @@ mapca: A Python implementation of the moving average principal components analys
 GIFT.
 """
 
-from .due import due, Doi
-
-from .info import (
-    __version__,
-    __author__,
-    __copyright__,
-    __credits__,
-    __license__,
-    __maintainer__,
-    __email__,
-    __status__,
-    __url__,
-    __packagename__,
-    __description__,
-    __longdesc__,
-)
-
-from .mapca import ma_pca
-
 import warnings
+
+from .due import Doi, due
+from .info import (__author__, __copyright__, __credits__, __description__,
+                   __email__, __license__, __longdesc__, __maintainer__,
+                   __packagename__, __status__, __url__, __version__)
+from .mapca import ma_pca
 
 # cmp is not used, so ignore nipype-generated warnings
 warnings.filterwarnings("ignore", r"cmp not installed")
