+++ mapca/mapca.py
@@ -3,7 +3,6 @@
 import logging
 
 import numpy as np
-
 from sklearn.decomposition import PCA
 from sklearn.preprocessing import StandardScaler
